reorganise with single entry point for applications
