+++ setup.cfg
@@ -45,10 +45,6 @@ install_requires =
     ruamel.yaml
 
 [options.extras_require]
-#Diagnostics =
-#    pandas
-#    matplotlib
-#    cartopy
 
 [options.packages.find]
 where =
@@ -59,23 +55,7 @@ where =
 
 [options.entry_points]
 console_scripts =
-    bkghandlingdriver.x = fv3jeditools.BkgHandling.BkgHandlingDriver:main
-    convertstatesgeos.x = fv3jeditools.Config.ConvertStatesGeos:main
-    processensemble.x = fv3jeditools.ConvertEnsemble.process_ensemble:main
-    processensemblegfsdiscover.x = fv3jeditools.ConvertEnsemble.process_ensemble_gfsdiscover:main
-    geosvar.x = fv3jeditools.DaRuns.GeosVar:main
-    femps_convergence.x = fv3jeditools.Diagnostics.femps_convergence:main
-    gradientnorm.x = fv3jeditools.Diagnostics.gradient_norm:main
-    hofx.x = fv3jeditools.Diagnostics.hofx:main
-    hofxomaomb.x = fv3jeditools.Diagnostics.hofx_pdf_omb_oma:main
-    twofieldrmscorr.x = fv3jeditools.Diagnostics.two_field_rms_corr:main
-    enshandlingdriver.x = fv3jeditools.EnsHandling.EnsHandlingDriver:main
-    geosens0hrfcst.x = fv3jeditools.EnsHandling.GeosEns0HourFcsts:main
-    obshandling_driver.x = fv3jeditools.ObsProcessing.ObsHandlingDriver:main
-    gsi_to_ioda.x = fv3jeditools.ObsProcessing.gsid_to_ioda:main
-    gsidiag_to_ioda.x = fv3jeditools.ObsProcessing.gsidiag_to_ioda:main
-    get_ensemble.x = fv3jeditools.EnsHandling.GetEnsemble:main
-    hpss_untar.x = fv3jeditools.FileHandling.hpss_untar:main
+    fv3jeditools.x = fv3jeditools.fv3jeditools:main
 
 [aliases]
 localdevelop = develop --user -e -b .
