--- src/fv3jeditools/FileHandling/out.yaml
@@ -1,1 +1,0 @@
-- ../../../Config/EMC/untar_ensemble.yml
