+++ src/fv3jeditools/Old/Diagnostics/gradient_norm.py
