+++ src/fv3jeditools/__init__.py
@@ -1,18 +1,6 @@
 # -*- coding: utf-8 -*-
 __path__ = __import__('pkgutil').extend_path(__path__, __name__)
 
-from .BkgHandling.BkgHandling import *
-
-from .Config.common_conf import *
-from .Config.geos_conf import *
-from .Config.gfs_conf import *
-
-from .ConvertEnsemble.fv3mod_ens_proc import *
-from .ConvertEnsemble.geos_ens_proc import *
-from .ConvertEnsemble.gfs_ens_proc import *
-
-from .EnsHandling.EnsHandling import *
-
-from .ObsProcessing.ObsHandling import *
-
-from .Utils.utils import *
+from .gsidiag_to_ioda import *
+from .untar import *
+from .utils import *
